feat: implementing grade reporting

+++ src/main/java/br/grupointegrado/trabalho/dto/NotaRequestDTO.java
@@ -1,5 +1,6 @@
 package br.grupointegrado.trabalho.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record NotaRequestDTO(Integer matriculaId, Integer disciplinaId, Double nota, LocalDate dataLancamento) {}
+public record NotaRequestDTO(Integer matriculaId, Integer disciplinaId, BigDecimal nota, LocalDate dataLancamento) {}
